Move LLM env/credentials into praisonai-code for C7 standalone gate.
Credential checks and endpoint resolution no longer import the wrapper; shims
preserve praisonai.llm.* module identity and CI smoke verifies standalone import.
