+++ src/praisonai-code/praisonai_code/llm/__init__.py
@@ -1,0 +1,12 @@
+"""LLM endpoint and credential helpers for the agentic CLI (wrapper-independent)."""
+
+from __future__ import annotations
+
+__all__ = [
+    "LLMEndpoint",
+    "default_model_for_available_provider",
+    "resolve_llm_endpoint",
+    "inject_credentials_into_env",
+    "is_configured",
+    "resolve_llm_endpoint_with_credentials",
+]
